requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
twilio>=7.0.0
websockets>=10.0
google-generativeai>=0.3.0
TTS>=0.8.0
python-multipart>=0.0.5
pydantic>=1.8.2
soundfile>=0.10.3
numpy>=1.21.0
aiohttp>=3.8.1 